+++ src/Sharpify.Data/CHANGELOG.md
@@ -15,6 +15,8 @@
     3. The `updateCondition` was not met
 * `Database` now tracks changes (additions, updates, removals) and compares them serialization events, to avoid serialization if no updates occurred since the previous serialization.
   * This means that you can automate serialization without worrying about potential waste of resources, for example you could omit `SerializeOnUpdate` from the `DatabaseConfiguration`, then create a background task that serializes on a given interval for example with `Sharpify.Routines.Routine` or `Sharpify.Routines.AsyncRoutine`, and it will only actually serialize if updates occurred. This can significantly improve performance in cases where there are write peaks, but the database is mostly read from.
+* You can now set the `Path` in the `DatabaseConfiguration` to an empty string `""` to receive an in-memory version of the database.
+It still has serialization methods, they don't perform any operations, they are essentially duds.
 
 ## v2.5.0